Move Style to new file and prep for css parsing. Update README.md
