+++ lib/content_element.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_html/html_elements.dart';
+import 'package:flutter_html/style.dart';
 import 'package:html/dom.dart' as dom;
 
 /// A [ContentElement] is a type of [StyledElement] that renders none of its [children].
