+++ lib/styled_element.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_html/style.dart';
 import 'package:html/dom.dart' as dom;
-
-import 'block_element.dart';
 
 /// A [StyledElement] applies a style to all of its children.
 class StyledElement {
@@ -151,64 +150,3 @@ StyledElement parseStyledElement(
 }
 
 typedef ListCharacter = String Function(int i);
-
-class Style {
-  TextStyle textStyle;
-  bool preserveWhitespace;
-  int baselineOffset;
-  String before;
-  String after;
-  TextDirection textDirection;
-  Block block;
-
-  Style({
-    this.textStyle,
-    this.preserveWhitespace,
-    this.baselineOffset,
-    this.before,
-    this.after,
-    this.textDirection,
-    this.block,
-  });
-
-  @override
-  String toString() {
-    return "(Text Style: ($textStyle}),)";
-  }
-
-  Style merge(Style other) {
-    if (other == null) return this;
-
-    Block mergedBlock = block?.merge(other.block);
-
-    return copyWith(
-      textStyle: other.textStyle,
-      preserveWhitespace: other.preserveWhitespace,
-      baselineOffset: other.baselineOffset,
-      before: other.before,
-      after: other.after,
-      textDirection: other.textDirection,
-      block: mergedBlock,
-    );
-  }
-
-  Style copyWith({
-    TextStyle textStyle,
-    bool preserveWhitespace,
-    int baselineOffset,
-    String before,
-    String after,
-    TextDirection textDirection,
-    Block block,
-  }) {
-    return Style(
-      textStyle: textStyle ?? this.textStyle,
-      preserveWhitespace: preserveWhitespace ?? this.preserveWhitespace,
-      baselineOffset: baselineOffset ?? this.baselineOffset,
-      before: before ?? this.before,
-      after: after ?? this.after,
-      textDirection: textDirection ?? this.textDirection,
-      block: block ?? this.block,
-    );
-  }
-}
